parseFloat()

+++ three.js
@@ -10,8 +10,8 @@ function getBiggest(a, b, c) {
 
 function onSubmit2() {
     event.preventDefault();
-    let a = parseInt(document.getElementById('input-3').value);
-    let b = parseInt(document.getElementById('input-4').value);
-    let c = parseInt(document.getElementById('input-5').value);
+    let a = parseFloat(document.getElementById('input-3').value);
+    let b = parseFloat(document.getElementById('input-4').value);
+    let c = parseFloat(document.getElementById('input-5').value);
     document.getElementById('output2').value = getBiggest(a, b, c);
 }
